+++ circus-train-aws-sns/src/main/java/com/hotels/bdp/circustrain/aws/sns/event/SnsListener.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,10 +15,11 @@
  */
 package com.hotels.bdp.circustrain.aws.sns.event;
 
-import java.io.UnsupportedEncodingException;
 import java.net.URI;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
@@ -42,7 +43,9 @@ import com.google.common.annotations.VisibleForTesting;
 import com.hotels.bdp.circustrain.api.CompletionCode;
 import com.hotels.bdp.circustrain.api.event.CopierListener;
 import com.hotels.bdp.circustrain.api.event.EventPartition;
+import com.hotels.bdp.circustrain.api.event.EventPartitions;
 import com.hotels.bdp.circustrain.api.event.EventReplicaCatalog;
+import com.hotels.bdp.circustrain.api.event.EventReplicaTable;
 import com.hotels.bdp.circustrain.api.event.EventSourceCatalog;
 import com.hotels.bdp.circustrain.api.event.EventTable;
 import com.hotels.bdp.circustrain.api.event.EventTableReplication;
@@ -72,6 +75,7 @@ public class SnsListener implements LocomotiveListener, SourceCatalogListener, R
   private String startTime;
   private EventSourceCatalog sourceCatalog;
   private EventReplicaCatalog replicaCatalog;
+  private LinkedHashMap<String, String> partitionKeyTypes;
 
   @Autowired
   public SnsListener(AmazonSNSAsyncClient sns, ListenerConfig config) {
@@ -103,18 +107,20 @@ public class SnsListener implements LocomotiveListener, SourceCatalogListener, R
   @Override
   public void tableReplicationStart(EventTableReplication tableReplication, String eventId) {
     startTime = clock.getTime();
-    SnsMessage message = new SnsMessage("START", config.getHeaders(), startTime, null, eventId, sourceCatalog.getName(),
-        replicaCatalog.getName(), tableReplication.getSourceTable().getQualifiedName(),
-        tableReplication.getQualifiedReplicaName(), null, null, null);
+    EventReplicaTable replicaTable = tableReplication.getReplicaTable();
+    SnsMessage message = new SnsMessage(SnsMessageType.START, config.getHeaders(), startTime, null, eventId, sourceCatalog.getName(),
+        replicaCatalog.getName(), replicaCatalog.getHiveMetastoreUris(), tableReplication.getSourceTable().getQualifiedName(),
+        tableReplication.getQualifiedReplicaName(), replicaTable.getTableLocation(), partitionKeyTypes, null, null, null);
     publish(config.getStartTopic(), message);
   }
 
   @Override
   public void tableReplicationSuccess(EventTableReplication tableReplication, String eventId) {
     String endTime = clock.getTime();
-    SnsMessage message = new SnsMessage("SUCCESS", config.getHeaders(), startTime, endTime, eventId,
-        sourceCatalog.getName(), replicaCatalog.getName(), tableReplication.getSourceTable().getQualifiedName(),
-        tableReplication.getQualifiedReplicaName(), getModifiedPartitions(partitionsToAlter, partitionsToCreate),
+    EventReplicaTable replicaTable = tableReplication.getReplicaTable();
+    SnsMessage message = new SnsMessage(SnsMessageType.SUCCESS, config.getHeaders(), startTime, endTime, eventId,
+        sourceCatalog.getName(), replicaCatalog.getName(), replicaCatalog.getHiveMetastoreUris(), tableReplication.getSourceTable().getQualifiedName(),
+        tableReplication.getQualifiedReplicaName(), replicaTable.getTableLocation(), partitionKeyTypes, getModifiedPartitions(partitionsToAlter, partitionsToCreate),
         getBytesReplicated(), null);
     publish(config.getSuccessTopic(), message);
   }
@@ -125,9 +131,10 @@ public class SnsListener implements LocomotiveListener, SourceCatalogListener, R
       startTime = clock.getTime();
     }
     String endTime = clock.getTime();
-    SnsMessage message = new SnsMessage("FAILURE", config.getHeaders(), startTime, endTime, eventId,
-        sourceCatalog.getName(), replicaCatalog.getName(), tableReplication.getSourceTable().getQualifiedName(),
-        tableReplication.getQualifiedReplicaName(), getModifiedPartitions(partitionsToAlter, partitionsToCreate),
+    EventReplicaTable replicaTable = tableReplication.getReplicaTable();
+    SnsMessage message = new SnsMessage(SnsMessageType.FAILURE, config.getHeaders(), startTime, endTime, eventId,
+        sourceCatalog.getName(), replicaCatalog.getName(), replicaCatalog.getHiveMetastoreUris(), tableReplication.getSourceTable().getQualifiedName(),
+        tableReplication.getQualifiedReplicaName(),replicaTable.getTableLocation(), partitionKeyTypes, getModifiedPartitions(partitionsToAlter, partitionsToCreate),
         getBytesReplicated(), t.getMessage());
     publish(config.getFailTopic(), message);
   }
@@ -140,15 +147,23 @@ public class SnsListener implements LocomotiveListener, SourceCatalogListener, R
   }
 
   @Override
-  public void partitionsToCreate(List<EventPartition> partitions) {
-    partitionsToCreate = partitions;
+  public void partitionsToCreate(EventPartitions eventPartitions) {
+    partitionsToCreate = eventPartitions.getEventPartitions();
+    setPartitionKeyTypes(eventPartitions.getPartitionKeyTypes());
   }
-
+  
   @Override
-  public void partitionsToAlter(List<EventPartition> partitions) {
-    partitionsToAlter = partitions;
+  public void partitionsToAlter(EventPartitions eventPartitions) {
+    partitionsToAlter = eventPartitions.getEventPartitions();
+    setPartitionKeyTypes(eventPartitions.getPartitionKeyTypes());
   }
-
+  
+  private void setPartitionKeyTypes(LinkedHashMap<String, String> partitionKeyTypes) {
+    if (partitionKeyTypes != null) {
+      this.partitionKeyTypes = partitionKeyTypes;
+    }
+  }
+  
   @VisibleForTesting
   static List<List<String>> getModifiedPartitions(
       List<EventPartition> partitionsToAlter,
@@ -171,8 +186,9 @@ public class SnsListener implements LocomotiveListener, SourceCatalogListener, R
     if (topic != null) {
       try {
         final String jsonMessage = startWriter.writeValueAsString(message);
-        if (jsonMessage.getBytes("UTF-8").length > SNS_MESSAGE_SIZE_LIMIT) {
-          LOG.warn("Message length execeeds SNS limit ({} bytes).", SNS_MESSAGE_SIZE_LIMIT);
+        int messageLength = jsonMessage.getBytes(StandardCharsets.UTF_8).length;
+        if (messageLength > SNS_MESSAGE_SIZE_LIMIT) {
+          LOG.warn("Message length of {} exceeds SNS limit ({} bytes).", messageLength, SNS_MESSAGE_SIZE_LIMIT);
         }
         LOG.debug("Attempting to send message to topic '{}': {}", topic, jsonMessage);
         PublishRequest request = new PublishRequest(topic, jsonMessage);
@@ -186,8 +202,6 @@ public class SnsListener implements LocomotiveListener, SourceCatalogListener, R
         }
       } catch (JsonProcessingException e) {
         LOG.error("Could not serialize message '{}'.", message, e);
-      } catch (UnsupportedEncodingException e) {
-        LOG.error("Could not UTF-8 encode message '{}'.", message, e);
       }
     }
   }
@@ -208,7 +222,7 @@ public class SnsListener implements LocomotiveListener, SourceCatalogListener, R
   public void resolvedReplicaLocation(URI location) {}
 
   @Override
-  public void existingReplicaPartitions(List<EventPartition> partitions) {}
+  public void existingReplicaPartitions(EventPartitions eventPartitions) {}
 
   @Override
   public void deprecatedReplicaLocations(List<URI> locations) {}
@@ -217,7 +231,7 @@ public class SnsListener implements LocomotiveListener, SourceCatalogListener, R
   public void resolvedMetaStoreSourceTable(EventTable table) {}
 
   @Override
-  public void resolvedSourcePartitions(List<EventPartition> partitions) {}
+  public void resolvedSourcePartitions(EventPartitions eventPartitions) {}
 
   @Override
   public void resolvedSourceLocation(URI location) {}