+++ circus-train-aws-sns/src/test/java/com/hotels/bdp/circustrain/aws/sns/event/SnsListenerTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,6 +26,7 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -41,7 +42,9 @@ import com.amazonaws.services.sns.AmazonSNSAsyncClient;
 import com.amazonaws.services.sns.model.PublishRequest;
 
 import com.hotels.bdp.circustrain.api.event.EventPartition;
+import com.hotels.bdp.circustrain.api.event.EventPartitions;
 import com.hotels.bdp.circustrain.api.event.EventReplicaCatalog;
+import com.hotels.bdp.circustrain.api.event.EventReplicaTable;
 import com.hotels.bdp.circustrain.api.event.EventSourceCatalog;
 import com.hotels.bdp.circustrain.api.event.EventSourceTable;
 import com.hotels.bdp.circustrain.api.event.EventTableReplication;
@@ -53,15 +56,19 @@ public class SnsListenerTest {
   private static final String SUBJECT = "choochoo";
   private static final RuntimeException ERROR = new RuntimeException("error message");
   private static final String ENDTIME = "endtime";
-  private static List<EventPartition> PARTITIONS_0;
-  private static List<EventPartition> PARTITIONS_1;
+  private static EventPartition PARTITION_0;
+  private static EventPartition PARTITION_1;
+
   private static final String STARTTIME = "starttime";
   private static final String EVENT_ID = "EVENT_ID";
+  private static final String PROTOCOL_VERSION = "1.1";
+  private static final String REPLICA_TABLE_LOCATION = "s3://bucket/path";
+  private static final String REPLICA_METASTORE_URIS = "thrift://host:9083";
 
   static {
     try {
-      PARTITIONS_0 = Arrays.asList(new EventPartition(Arrays.asList("2014-01-01", "0"), new URI("location_0")));
-      PARTITIONS_1 = Arrays.asList(new EventPartition(Arrays.asList("2014-01-01", "1"), new URI("location_1")));
+      PARTITION_0 = new EventPartition(Arrays.asList("2014-01-01", "0"), new URI("location_0"));
+      PARTITION_1 = new EventPartition(Arrays.asList("2014-01-01", "1"), new URI("location_1"));
     } catch (URISyntaxException e) {}
   }
   @Mock
@@ -75,6 +82,8 @@ public class SnsListenerTest {
   @Mock
   private EventSourceTable sourceTable;
   @Mock
+  private EventReplicaTable replicaTable;
+  @Mock
   private EventTableReplication tableReplication;
   @Mock
   private Metrics metrics;
@@ -83,9 +92,14 @@ public class SnsListenerTest {
   @Captor
   private ArgumentCaptor<PublishRequest> requestCaptor;
 
+  private LinkedHashMap<String, String> partitionKeyTypes = new LinkedHashMap<>();
+
   @Before
   public void prepare() throws URISyntaxException {
     when(clock.getTime()).thenReturn(STARTTIME, ENDTIME);
+
+    partitionKeyTypes.put("local_date", "string");
+    partitionKeyTypes.put("local_hour", "int");
 
     Map<String, String> headers = new HashMap<>();
     headers.put("pipeline-id", "0943879438");
@@ -97,11 +111,14 @@ public class SnsListenerTest {
     when(config.getHeaders()).thenReturn(headers);
     when(config.getQueueSize()).thenReturn(10);
 
+    when(replicaTable.getTableLocation()).thenReturn(REPLICA_TABLE_LOCATION);
     when(tableReplication.getSourceTable()).thenReturn(sourceTable);
+    when(tableReplication.getReplicaTable()).thenReturn(replicaTable);
     when(tableReplication.getQualifiedReplicaName()).thenReturn("replicaDb.replicaTable");
     when(metrics.getBytesReplicated()).thenReturn(40L);
     when(sourceCatalog.getName()).thenReturn("sourceCatalogName");
     when(replicaCatalog.getName()).thenReturn("replicaCatalogName");
+    when(replicaCatalog.getHiveMetastoreUris()).thenReturn(REPLICA_METASTORE_URIS);
     when(sourceTable.getQualifiedName()).thenReturn("srcDb.srcTable");
   }
 
@@ -115,20 +132,27 @@ public class SnsListenerTest {
     PublishRequest request = requestCaptor.getValue();
     assertThat(request.getSubject(), is(SUBJECT));
     assertThat(request.getTopicArn(), is("startArn"));
-    assertThat(request.getMessage(),
-        is("{\"protocolVersion\":\"1.0\",\"type\":\"START\",\"headers\":{\"pipeline-id\":\"0943879438\"},"
-            + "\"startTime\":\"starttime\",\"eventId\":\"EVENT_ID\",\"sourceCatalog\":\"sourceCatalogName\","
-            + "\"replicaCatalog\":\"replicaCatalogName\",\"sourceTable\":\"srcDb.srcTable\",\"replicaTable\":"
-            + "\"replicaDb.replicaTable\"}"));
+    assertThat(request.getMessage(), is("{\"protocolVersion\":\"" + PROTOCOL_VERSION + "\"" 
+        + ",\"type\":\"START\",\"headers\":{\"pipeline-id\":\"0943879438\"},"
+        + "\"startTime\":\"starttime\",\"eventId\":\"EVENT_ID\",\"sourceCatalog\":\"sourceCatalogName\","
+        + "\"replicaCatalog\":\"replicaCatalogName\",\"sourceTable\":\"srcDb.srcTable\",\"replicaTable\":"
+        + "\"replicaDb.replicaTable\",\"replicaTableLocation\":\"s3://bucket/path\",\"replicaMetastoreUris\":\"thrift://host:9083\"}"));
   }
 
   @Test
-  public void success() {
+  public void successPartitionedTable() {
     SnsListener listener = new SnsListener(client, config, clock);
     listener.circusTrainStartUp(new String[] {}, sourceCatalog, replicaCatalog);
     listener.tableReplicationStart(tableReplication, EVENT_ID);
-    listener.partitionsToAlter(PARTITIONS_0);
-    listener.partitionsToCreate(PARTITIONS_1);
+
+    EventPartitions alteredPartitions = new EventPartitions(partitionKeyTypes);
+    alteredPartitions.add(PARTITION_0);
+    listener.partitionsToAlter(alteredPartitions);
+
+    EventPartitions createdPartitions = new EventPartitions(partitionKeyTypes);
+    createdPartitions.add(PARTITION_1);
+    listener.partitionsToCreate(createdPartitions);
+
     listener.copierEnd(metrics);
     listener.tableReplicationSuccess(tableReplication, EVENT_ID);
 
@@ -136,12 +160,43 @@ public class SnsListenerTest {
     PublishRequest request = requestCaptor.getAllValues().get(1);
     assertThat(request.getSubject(), is(SUBJECT));
     assertThat(request.getTopicArn(), is("successArn"));
-    assertThat(request.getMessage(),
-        is("{\"protocolVersion\":\"1.0\",\"type\":\"SUCCESS\",\"headers\":{\"pipeline-id\":\"0943879438\"},"
+    assertThat(request.getMessage(), is("{\"protocolVersion\":\"" + PROTOCOL_VERSION + "\""
+            + ",\"type\":\"SUCCESS\",\"headers\":{\"pipeline-id\":\"0943879438\"},"
             + "\"startTime\":\"starttime\",\"endTime\":\"endtime\",\"eventId\":\"EVENT_ID\",\"sourceCatalog\""
             + ":\"sourceCatalogName\",\"replicaCatalog\":\"replicaCatalogName\",\"sourceTable\":"
-            + "\"srcDb.srcTable\",\"replicaTable\":\"replicaDb.replicaTable\",\"modifiedPartitions\":"
+            + "\"srcDb.srcTable\",\"replicaTable\":\"replicaDb.replicaTable\","
+            + "\"replicaTableLocation\":\""
+            + REPLICA_TABLE_LOCATION
+            + "\",\"replicaMetastoreUris\":\""
+            + REPLICA_METASTORE_URIS
+            + "\",\"partitionKeys\":{\"local_date\":\"string\",\"local_hour\":\"int\"},"
+            + "\"modifiedPartitions\":"
             + "[[\"2014-01-01\",\"0\"],[\"2014-01-01\",\"1\"]],\"bytesReplicated\":40}"));
+  }
+
+  @Test
+  public void successUnpartitionedTable() {
+    SnsListener listener = new SnsListener(client, config, clock);
+    listener.circusTrainStartUp(new String[] {}, sourceCatalog, replicaCatalog);
+    listener.tableReplicationStart(tableReplication, EVENT_ID);
+
+    listener.copierEnd(metrics);
+    listener.tableReplicationSuccess(tableReplication, EVENT_ID);
+
+    verify(client, times(2)).publish(requestCaptor.capture());
+    PublishRequest request = requestCaptor.getAllValues().get(1);
+    assertThat(request.getSubject(), is(SUBJECT));
+    assertThat(request.getTopicArn(), is("successArn"));
+    assertThat(request.getMessage(), is("{\"protocolVersion\":\"" + PROTOCOL_VERSION + "\""
+            + ",\"type\":\"SUCCESS\",\"headers\":{\"pipeline-id\":\"0943879438\"},"
+            + "\"startTime\":\"starttime\",\"endTime\":\"endtime\",\"eventId\":\"EVENT_ID\",\"sourceCatalog\""
+            + ":\"sourceCatalogName\",\"replicaCatalog\":\"replicaCatalogName\",\"sourceTable\":"
+            + "\"srcDb.srcTable\",\"replicaTable\":\"replicaDb.replicaTable\","
+            + "\"replicaTableLocation\":\""
+            + REPLICA_TABLE_LOCATION
+            + "\",\"replicaMetastoreUris\":\""
+            + REPLICA_METASTORE_URIS
+            + "\",\"bytesReplicated\":40}"));
   }
 
   @Test
@@ -149,8 +204,15 @@ public class SnsListenerTest {
     SnsListener listener = new SnsListener(client, config, clock);
     listener.circusTrainStartUp(new String[] {}, sourceCatalog, replicaCatalog);
     listener.tableReplicationStart(tableReplication, EVENT_ID);
-    listener.partitionsToAlter(PARTITIONS_0);
-    listener.partitionsToCreate(PARTITIONS_1);
+
+    EventPartitions alteredPartitions = new EventPartitions(partitionKeyTypes);
+    alteredPartitions.add(PARTITION_0);
+    listener.partitionsToAlter(alteredPartitions);
+
+    EventPartitions createdPartitions = new EventPartitions(partitionKeyTypes);
+    createdPartitions.add(PARTITION_1);
+    listener.partitionsToCreate(createdPartitions);
+
     listener.copierEnd(metrics);
     listener.tableReplicationFailure(tableReplication, EVENT_ID, ERROR);
 
@@ -158,11 +220,17 @@ public class SnsListenerTest {
     PublishRequest request = requestCaptor.getValue();
     assertThat(request.getSubject(), is(SUBJECT));
     assertThat(request.getTopicArn(), is("failArn"));
-    assertThat(request.getMessage(), is("{\"protocolVersion\":\"1.0\",\"type\":\"FAILURE\",\"headers\":"
+    assertThat(request.getMessage(), is("{\"protocolVersion\":\"" + PROTOCOL_VERSION + "\""
+        + ",\"type\":\"FAILURE\",\"headers\":"
         + "{\"pipeline-id\":\"0943879438\"},\"startTime\":\"starttime\",\"endTime\":\"endtime\",\"eventId\":"
         + "\"EVENT_ID\",\"sourceCatalog\":\"sourceCatalogName\",\"replicaCatalog\":\"replicaCatalogName\","
-        + "\"sourceTable\":\"srcDb.srcTable\",\"replicaTable\":\"replicaDb.replicaTable\",\"modifiedPartitions\":"
-        + "[[\"2014-01-01\",\"0\"],[\"2014-01-01\",\"1\"]],\"bytesReplicated\":40,\"errorMessage\":\"error message\"}"));
+        + "\"sourceTable\":\"srcDb.srcTable\",\"replicaTable\":\"replicaDb.replicaTable\","
+        + "\"replicaTableLocation\":\""
+        + REPLICA_TABLE_LOCATION
+        + "\",\"replicaMetastoreUris\":\""
+        + REPLICA_METASTORE_URIS
+        + "\",\"partitionKeys\":{\"local_date\":\"string\",\"local_hour\":\"int\"},"
+        + "\"modifiedPartitions\":[[\"2014-01-01\",\"0\"],[\"2014-01-01\",\"1\"]],\"bytesReplicated\":40,\"errorMessage\":\"error message\"}"));
   }
 
   @Test
@@ -175,10 +243,11 @@ public class SnsListenerTest {
     PublishRequest request = requestCaptor.getValue();
     assertThat(request.getSubject(), is(SUBJECT));
     assertThat(request.getTopicArn(), is("failArn"));
-    assertThat(request.getMessage(), is("{\"protocolVersion\":\"1.0\",\"type\":\"FAILURE\",\"headers\":"
+    assertThat(request.getMessage(), is("{\"protocolVersion\":\"" + PROTOCOL_VERSION + "\""
+        + ",\"type\":\"FAILURE\",\"headers\":"
         + "{\"pipeline-id\":\"0943879438\"},\"startTime\":\"starttime\",\"endTime\":\"endtime\",\"eventId\":"
         + "\"EVENT_ID\",\"sourceCatalog\":\"sourceCatalogName\",\"replicaCatalog\":\"replicaCatalogName\","
-        + "\"sourceTable\":\"srcDb.srcTable\",\"replicaTable\":\"replicaDb.replicaTable\",\"bytesReplicated\":0,\"errorMessage\":\"error message\"}"));
+        + "\"sourceTable\":\"srcDb.srcTable\",\"replicaTable\":\"replicaDb.replicaTable\",\"replicaTableLocation\":\"s3://bucket/path\",\"replicaMetastoreUris\":\"thrift://host:9083\",\"bytesReplicated\":0,\"errorMessage\":\"error message\"}"));
   }
 
   @Test