+++ circus-train-core/src/test/java/com/hotels/bdp/circustrain/core/PartitionsAndStatisticsTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -66,6 +66,7 @@ public class PartitionsAndStatisticsTest {
 
     assertThat(partitionsAndStatistics.getPartitionNames(), is(expectedName));
     assertThat(partitionsAndStatistics.getPartitions(), is(partitions));
+    assertThat(partitionsAndStatistics.getPartitionKeys(), is(partitionKeys));
     ColumnStatisticsDesc statsDesc = new ColumnStatisticsDesc(false, "db1", "t1");
     statsDesc.setPartName("a=b/c=d");
     ColumnStatistics expectedStats = new ColumnStatistics(statsDesc, columnStats);
@@ -82,6 +83,7 @@ public class PartitionsAndStatisticsTest {
 
     assertThat(partitionsAndStatistics.getPartitionNames(), is(empty()));
     assertThat(partitionsAndStatistics.getPartitions(), is(empty()));
+    assertThat(partitionsAndStatistics.getPartitionKeys(), is(partitionKeys));
   }
 
   @Test(expected = IllegalArgumentException.class)