+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
+### Added
+* Added `replicaTableLocation`, `replicaMetastoreUris` and  `partitionKeys` to SNS message.
 ### Changed
+* SNS Message `protocolVersion` changed from "1.0" to "1.1".
+* Updated documentation for circus-train-aws-sns module (full reference of SNS message format, more examples).
 * Fixed references to README.md in command line runner help messages to point to correct GitHub locations.
 
 # 11.0.0 - 2018-01-16